group create request action return
